requirements.txt
psutil>=5.9.0
pystray>=0.19.0
Pillow>=9.0.0
streamlit>=1.30.0
pyinstaller>=6.0.0
solana>=0.30.0
solders>=0.20.0
websockets==15.0
aiohttp>=3.9.0